backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib
argon2-cffi
python-socketio
pgvector
ollama
pypdf
httpx
razorpay
stripe

# Document Processing & OCR
pytesseract==0.3.10
Pillow==10.0.0
pdf2image==1.16.3
pdfplumber==0.10.3
openpyxl==3.1.2
pandas==2.0.3
python-pptx==0.6.21
python-dotenv
groq
sentence-transformers